+++ account-kit/infra/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.88.2](https://github.com/alchemyplatform/aa-sdk/compare/v4.88.1...v4.88.2) (2026-05-06)
+
+**Note:** Version bump only for package @account-kit/infra
+
 ## [4.88.1](https://github.com/alchemyplatform/aa-sdk/compare/v4.88.0...v4.88.1) (2026-04-24)
 
 **Note:** Version bump only for package @account-kit/infra